requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
numpy
opencv-python
soundfile
librosa
openai-whisper
deepface
python-multipart
pytest
sounddevice
httpx>=0.24,<1.0